Rename error_codes -> return_codes
