+++ collision_sensor/include/collision_sensor/contact_detection/return_codes.hpp
